updated extractor contribution policy
Updated extractor contribution policy. This is because
extractors has a funny way of behaving. They can pass a
test to day and fail the next day. So to make sure out
master branch is always clean and pass it build, all
extractor tests has been disabled in the master branch.
It is only extractors that passes it tests in the extractor_testing branch will be added to the master
branch